Added JSON reporter
